# apps/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "PlaybackUrlResponse": {
        "additionalProperties": false,
        "properties": {
          "expires_in_seconds": {
            "default": 300,
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "playback_url": {
            "title": "Playback Url",
            "type": "string"
          },
          "recording_id": {
            "format": "uuid",
            "title": "Recording Id",
            "type": "string"
          }
        },
        "required": [
          "recording_id",
          "playback_url"
        ],
        "title": "PlaybackUrlResponse",
        "type": "object"
      },
...
      "RecordingCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "assignment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Id"
          },
          "duration_seconds": {
            "maximum": 3600.0,
            "minimum": 300.0,
            "title": "Duration Seconds",
            "type": "integer"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "studio_id": {
            "format": "uuid",
            "title": "Studio Id",
            "type": "string"
          }
        },
        "required": [
          "studio_id",
          "mime_type",
          "duration_seconds"
        ],
        "title": "RecordingCreateRequest",
        "type": "object"
      },
      "RecordingResponse": {
        "additionalProperties": false,
        "properties": {
          "assignment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "integer"
          },
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "r2_object_key": {
            "title": "R2 Object Key",
            "type": "string"
          },
          "student_id": {
            "format": "uuid",
            "title": "Student Id",
            "type": "string"
          },
          "studio_id": {
            "format": "uuid",
            "title": "Studio Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "uploaded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded At"
          }
        },
        "required": [
          "id",
          "studio_id",
          "student_id",
          "r2_object_key",
          "mime_type",
          "duration_seconds",
          "created_at",
          "updated_at"
        ],
        "title": "RecordingResponse",
        "type": "object"
      },
...
      "UploadConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "file_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Size Bytes"
          }
        },
        "title": "UploadConfirmRequest",
        "type": "object"
      },
      "UploadUrlResponse": {
        "additionalProperties": false,
        "properties": {
          "r2_object_key": {
            "title": "R2 Object Key",
            "type": "string"
          },
          "recording_id": {
            "format": "uuid",
            "title": "Recording Id",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "recording_id",
          "upload_url",
          "r2_object_key"
        ],
        "title": "UploadUrlResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DailyRiff API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/recordings": {
      "get": {
        "operationId": "list_recordings_recordings_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecordingResponse"
                  },
                  "title": "Response List Recordings Recordings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Recordings",
        "tags": [
          "recordings"
        ]
      },
      "post": {
        "operationId": "create_recording_recordings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Recording",
        "tags": [
          "recordings"
        ]
      }
    },
    "/recordings/{recording_id}": {
      "get": {
        "operationId": "get_recording_recordings__recording_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Recording not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Recording",
        "tags": [
          "recordings"
        ]
      }
    },
    "/recordings/{recording_id}/confirm-upload": {
      "post": {
        "description": "Confirm that the client has completed uploading to R2.\n\nSets uploaded_at which triggers the auto_acknowledge_assignment DB trigger.",
        "operationId": "confirm_upload_recordings__recording_id__confirm_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Recording not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Upload",
        "tags": [
          "recordings"
        ]
      }
    },
    "/recordings/{recording_id}/playback-url": {
      "get": {
        "description": "Policy check \u2192 mint 5-min presigned R2 URL. FastAPI never proxies bytes.",
        "operationId": "get_playback_url_recordings__recording_id__playback_url_get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackUrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized to play this recording"
          },
          "404": {
            "description": "Recording not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Playback Url",
        "tags": [
          "recordings"
        ]
      }
    },
...
  }
}